build: make sedutil-cli binary static
some of the more ancient rootfs-es we've still got in circulation have
versions of `libstdc++.so` that are even older than the one in our
`lb-build` container. e.g. rootfs.PRODUCT_BASE_CENTOS refuses to run
dynamically linked `sedutil-cli` with:
    /tmp/sedutil-cli: /lib64/libstdc++.so.6: version `GLIBCXX_3.4.26'
            not found (required by /tmp/sedutil-cli)
    /tmp/sedutil-cli: /lib64/libstdc++.so.6: version `GLIBCXX_3.4.21'
            not found (required by /tmp/sedutil-cli)

build sedutil-cli as a static binary instead, so it can be used on most
relevant x86-64 rootfs-es.